Transferred needed logic to Service layer.
Most logic is now is Service layer. Kept the convertions to DTO in Controller layer because thought it was more appropriate there. Fixed some minor things here and there

+++ src/main/java/com/logistics/logisticsCompany/repository/OfficeRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface OfficeRepository extends JpaRepository<Office, Integer> {
+public interface OfficeRepository extends JpaRepository<Office, Long> {
     boolean existsByOfficeName(String officeName);
     boolean existsByAddress(String address);
     boolean existsById(long id);
